pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scribeflow"
version = "0.1.0"
description = "ScribeFlow discovery layer and visual manifest service"
requires-python = ">=3.10"
dependencies = [
  "markitdown>=0.1.1",
  "openai>=1.40.0",
  "python-dotenv>=1.0.1",
]

[project.scripts]
scribeflow = "scribeflow.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
